get server tests working and fix timeout issues

+++ server/routes/medias-routes.js
@@ -43,20 +43,20 @@ const {
 let upload = (req, res, next) => {
   //console.log('upload',req, res);
   multerUploadSingleFile(req, res, function (err) {
-    console.log('media-routes multerUploadSingleFile', err);
-    console.log('req.body.media', req.body.media);
+    //console.log('media-routes multerUploadSingleFile', err);
+    //console.log('req.body.media', req.body.media);
     req.passedMedia = JSON.parse(req.body.media);
     delete req.body.media;
-    console.log('req.file', req.file);
+    //console.log('req.file', req.file);
     if (err) {
       processErr(err);
     } else {
       if (!req.file) {
-        console.log('No file was selected');
+        //console.log('No file was selected');
       } else {
-        console.log('media patch File uploaded!');
+        //console.log('media patch File uploaded!');
         let fileName = req.file.filename;
-        console.log("newFileName", fileName);
+        //console.log("newFileName", fileName);
         let location = path.join(req.file.destination, fileName);
         req.passedMedia.location = location;
         req.passedMedia.originalFileName = req.file.originalname;
@@ -78,40 +78,44 @@ let addUserToComments = (media) => {
     let processedComments = 0;
     let commentsArr = [];
 
-    console.log('addUserToComments', 'numComments', numComments);
-    for (let comment of retMedia.comments) {
-      console.log('addUserToComments', 'comment', comment);
-      let userObj = {
-        '_creatorRef': comment._creator
+    //console.log('addUserToComments', 'numComments', numComments);
+    if (retMedia.comments && retMedia.comments.length > 0) {
+      for (let comment of retMedia.comments) {
+        //console.log('addUserToComments', 'comment', comment);
+        let userObj = {
+          '_creatorRef': comment._creator
+        };
+        User.findOne(userObj).populate('_profileMediaId', ['location']).then((user) => {
+          //console.log('addUserToComments', 'user', user);
+          //console.log('addUserToComments', 'typeof user', typeof user);
+          //console.log('addUserToComments', 'typeof comment', typeof comment);
+          if (user) {
+            delete user._id;
+            let newComment = JSON.parse(JSON.stringify(comment));
+            //console.log('addUserToComments', 'user.name', user.name);
+            //console.log('addUserToComments', 'user._profileMediaId', user._profileMediaId);
+            newComment.user = {};
+            //console.log('addUserToComments', 'newComment.user', newComment.user);
+            newComment.user.name = user.name;
+            newComment.user._profileMediaId = user._profileMediaId;
+            //console.log('addUserToComments', 'newComment.user2', newComment.user);
+
+            commentsArr.push(newComment);
+            //console.log('addUserToComments', 'newComment', newComment);
+          };
+
+          processedComments++;
+          if (numComments === processedComments) {
+            retMedia.comments = commentsArr;
+            //console.log('addUserToComments', 'resolve', retMedia);
+            return resolve(retMedia);
+          };
+        }, (e) => {
+          reject(e);
+        });
       };
-      User.findOne(userObj).populate('_profileMediaId', ['location']).then((user) => {
-        console.log('addUserToComments', 'user', user);
-        console.log('addUserToComments', 'typeof user', typeof user);
-        console.log('addUserToComments', 'typeof comment', typeof comment);
-        if (user) {
-          delete user._id;
-          let newComment = JSON.parse(JSON.stringify(comment));
-          console.log('addUserToComments', 'user.name', user.name);
-          console.log('addUserToComments', 'user._profileMediaId', user._profileMediaId);
-          newComment.user = {};
-          console.log('addUserToComments', 'newComment.user', newComment.user);
-          newComment.user.name = user.name;
-          newComment.user._profileMediaId = user._profileMediaId;
-          console.log('addUserToComments', 'newComment.user2', newComment.user);
-
-          commentsArr.push(newComment);
-          console.log('addUserToComments', 'newComment', newComment);
-        };
-
-        processedComments++;
-        if (numComments === processedComments) {
-          retMedia.comments = commentsArr;
-          console.log('addUserToComments', 'resolve', retMedia);
-          return resolve(retMedia);
-        };
-      }, (e) => {
-        reject(e);
-      });
+    } else {
+      return resolve(retMedia);
     };
   });
 };
@@ -123,38 +127,42 @@ let transformCreatorToUser = (medias) => {
     let numMedias = medias ? medias.length : 0;
     let processedMedias = 0;
     let transformedMedias = [];
-    for (let media of medias) {
-      if (media.comments && media.comments.length > 0) {
-        let numComments = media.comments.length;
-        let processedComments = 0;
-        for (let comment of media.comments) {
-          let userObj = {
-            '_creatorRef': comment._creator
-          };
-          User.findOne(userObj).populate('_profileMediaId', ['location']).then((user) => {
-            if (user) {
-              delete user._id;
-              comment.user = user;
+    if (numMedias > 0) {
+      for (let media of medias) {
+        if (media.comments && media.comments.length > 0) {
+          let numComments = media.comments.length;
+          let processedComments = 0;
+          for (let comment of media.comments) {
+            let userObj = {
+              '_creatorRef': comment._creator
             };
-            processedComments++;
-            if (numComments === processedComments) {
-              transformedMedias.push(media);
-              processedMedias++;
-              if (numMedias === processedMedias) {
-                return resolve(transformedMedias);
+            User.findOne(userObj).populate('_profileMediaId', ['location']).then((user) => {
+              if (user) {
+                delete user._id;
+                comment.user = user;
               };
-            };
-          }, (e) => {
-            reject(e);
-          });
-        };
-      } else {
-        transformedMedias.push(media);
-        processedMedias++;
-        if (numMedias === processedMedias) {
-          return resolve(transformedMedias);
+              processedComments++;
+              if (numComments === processedComments) {
+                transformedMedias.push(media);
+                processedMedias++;
+                if (numMedias === processedMedias) {
+                  return resolve(transformedMedias);
+                };
+              };
+            }, (e) => {
+              reject(e);
+            });
+          };
+        } else {
+          transformedMedias.push(media);
+          processedMedias++;
+          if (numMedias === processedMedias) {
+            return resolve(transformedMedias);
+          };
         };
       };
+    } else {
+      return resolve(transformedMedias);
     };
   });
 };
@@ -163,12 +171,16 @@ let downloadFiles = (medias) => {
   return new Promise((resolve, reject) => {
     let numMedias = medias ? medias.length : 0;
     let processedMedias = 0;
-    for (let media of medias) {
-      downloadFile(media);
-      processedMedias++;
-      if (numMedias === processedMedias) {
-        return resolve();
-      };
+    if (numMedias > 0) {
+      for (let media of medias) {
+        downloadFile(media);
+        processedMedias++;
+        if (numMedias === processedMedias) {
+          return resolve();
+        };
+      }
+    } else {
+      return resolve();
     };
   });
 };
@@ -190,9 +202,9 @@ let downloadFile = (media) => {
 router.post('/', authenticate, upload, (req, res) => {
   if (!req.loggedInUser.guestUser) {
     let body = _.pick(req.passedMedia, mediaInsertFields);
-    console.log('body', body);
+    //console.log('body', body);
     let media = new Media(body);
-    console.log('body.photoInfo', body.photoInfo);
+    //console.log('body.photoInfo', body.photoInfo);
     if (body.photoInfo && body.photoInfo.location && body.photoInfo.isUrl) {
       media.location = body.photoInfo.location;
       media.isUrl = true;
@@ -204,10 +216,10 @@ router.post('/', authenticate, upload, (req, res) => {
     media.addedDate = new Date().getTime();
     media.comments = [];
     media._id = new ObjectID();
-    console.log('media', media);
+    //console.log('media', media);
 
     media.save().then((newMedia) => {
-      console.log('media2', newMedia);
+      //console.log('media2', newMedia);
       res.send(_.pick(newMedia, mediaOutFields));
     }, (e) => {
       console.log('media.save() e', e);
@@ -254,6 +266,7 @@ router.get('/byCriteria', authenticate, (req, res) => {
 
   Media.findByCriteria(tags, people, fromDate, toDate).then((medias) => {
     let obj = {};
+    obj['medias'] = medias;
     downloadFiles(medias).then(() => {
       res.send(obj);
     }, (e) => {
@@ -270,9 +283,9 @@ router.get('/title/:title', authenticate, (req, res) => {
   let mediaObj = {
     title
   };
-  console.log("mediaObj", mediaObj);
-  Media.findOne(mediaObj).populate('comments tags people').then((media) => {
-    console.log("media", media);
+  //console.log("mediaObj", mediaObj);
+  Media.findOne(mediaObj).then((media) => {
+    //console.log("media", media);
     if (media) {
       res.send({
         'titleFound': true,
@@ -348,7 +361,10 @@ router.delete('/:id', authenticate, (req, res) => {
 
       if (media) {
         googleCloudApi.deleteFile(media.location);
-        fs.unlinkSync(media.location);
+        if (fs.existsSync(media.location)) {
+          fs.unlinkSync(media.location);
+        };
+
         res.send({
           media
         });
@@ -399,7 +415,7 @@ router.delete('/:id', authenticate, (req, res) => {
 
 let updateMedias = (res, body, medias, commentId) => {
 
-  console.log("updateMedias", body, medias, commentId);
+  //console.log("updateMedias", body, medias, commentId);
 
   // let mediasObj = {
   //   _id : medias._id
@@ -414,8 +430,8 @@ let updateMedias = (res, body, medias, commentId) => {
       "comments": commentId
     };
   };
-  console.log("updateObj", updateObj);
-  console.log("medias", medias);
+  //console.log("updateObj", updateObj);
+  //console.log("medias", medias);
 
   Media.findOneAndUpdate(medias, updateObj, {
     new: true
@@ -428,7 +444,7 @@ let updateMedias = (res, body, medias, commentId) => {
           });
         }, (e) => {
           res.status(400).send();
-          console.log(e);
+          //console.log(e);
         });
       } else {
         res.send({
@@ -443,13 +459,13 @@ let updateMedias = (res, body, medias, commentId) => {
 
   }, (e) => {
     res.status(400).send();
-    console.log(e);
+    //console.log(e);
   });
 };
 
 
 router.patch('/:id', authenticate, upload, (req, res) => {
-  console.log("router.patch1", req.passedMedia);
+  //console.log("router.patch1", req.passedMedia);
   if (!req.loggedInUser.guestUser) {
     let {
       id
@@ -457,7 +473,7 @@ router.patch('/:id', authenticate, upload, (req, res) => {
 
     let body = _.pick(req.passedMedia, mediaUpdateFields);
 
-    console.log("router.patch2", body);
+    //console.log("router.patch2", body);
 
     if (!ObjectID.isValid(id)) {
       return res.status(404).send({
@@ -479,7 +495,7 @@ router.patch('/:id', authenticate, upload, (req, res) => {
 
       comment._creator = req.loggedInUser._creatorRef;
       comment.commentDate = new Date().getTime();
-      console.log('comment', comment);
+      //console.log('comment', comment);
 
       comment.save().then((comment) => {
         updateMedias(res, body, medias, comment._id);
